+++ .gitignore
@@ -5,3 +5,4 @@ build/
 dist/
 *.egg-info/
 docs/_build/
+venv
