+++ requirements.txt
@@ -1,2 +1,3 @@
 boto
 requests
+redis
